refactor(pipeline): rename types_max/assemble to project_seed/project_render
Both modules lost their max-depth contents when that generation depth was
removed; only ProjectOut and render_project (PR #990's deterministic
Projects normalization) survive. Rename the files to name what they
actually hold and trim the apologetic why-the-filename-is-wrong headers
down to their actual contract; that history belongs in the ADR.
