+++ apps/desktop/src-tauri/src/pipeline/resume/project_seed.rs
@@ -1,15 +1,8 @@
 //! The Projects seed type shared by the deterministic quality-depth Projects
 //! normalization (`pipeline::resume::projects`) and its seeder
-//! (`pipeline::resume::source`).
-//!
-//! This module used to carry the whole max-depth section-generation type
-//! family (one typed answer per résumé section, plus the max-only judge,
-//! shown to a model over JSON — which is why every type there was flat and
-//! serde-round-trippable). That machinery was removed with the `max`
-//! generation depth; [`ProjectOut`] survives as a plain in-process value —
-//! it is also the seed type PR #990's deterministic Projects normalization
-//! renders through `assemble::render_project`, unrelated to depth, and
-//! nothing (de)serializes it any more.
+//! (`pipeline::resume::source`). [`ProjectOut`] is a plain in-process value —
+//! nothing (de)serializes it — rendered through
+//! `project_render::render_project`.
 
 /// ONE project entry, in the owner-locked source signature.
 ///
